Moved tests to respective folders
